Replaced track following code in TrolleyWaypointController with SplineScript.cs; removed old track implementation from scene
