--- Assets/00_Custom/Scripts/SplineScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c8810570eb037284983e2c85493b6afe
